Updated js to tsx, added data interfaces
