--- pantry/src/pages/grocery_list.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const GroceryList = () => {
-return (
-	<h1>GROCERY LIST!!!!</h1>
-);
-};
-
-export default GroceryList;
