+++ pantry/src/pages/grocery_list.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+const GroceryList = () => {
+  return <h1>GROCERY LIST!!!!</h1>;
+};
+export default GroceryList;
