--- pantry/src/pages/index.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Home = () => {
-return (
-	<div>
-	<h1>Temp???</h1>
-	</div>
-);
-};
-
-export default Home;
