+++ pantry/src/pages/index.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Home = () => {
+  return (
+    <div>
+      <h1>Temp???</h1>
+    </div>
+  );
+};
+
+export default Home;
