--- pantry/src/pages/meal_plan.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const MealPlan = () => {
-return (
-	<h1>MEAL PLAN!!!!</h1>
-);
-};
-
-export default MealPlan;
