+++ pantry/src/pages/meal_plan.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+const MealPlan = () => {
+  return <h1>MEAL PLAN!!!!</h1>;
+};
+export default MealPlan;
