--- pantry/src/pages/recipes.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Recipes = () => {
-return (
-	<h1>Recipes!!!!</h1>
-);
-};
-
-export default Recipes;
