+++ pantry/src/pages/recipes.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Recipes = () => {
+  return <h1>Recipes!!!!</h1>;
+};
+
+export default Recipes;
